refactor(activerecord): relocate test-helper concerns out of test-adapter.ts
Mechanical move of three independent test-helper concerns from
test-adapter.ts (769 LOC) into dedicated files under test-helpers/.
Zero semantic change — same module-level state, same public surface,
just moved closer to consumers.

New files:
- test-helpers/ddl-tracker.ts — _createdTables/_createdColumns state,
  parseCreateTableColumns SQL DDL parser, recordDdlTracking, and the
  snapshot/restore pair used by withTransactionalFixtures. Dropped the
  underscore from snapshot/restore now that they're module-public.
- test-helpers/skip-global-reset.ts — withTransactionalFixtures scope
  refcount (push/pop/shouldSkipGlobalReset).
- test-helpers/use-transactional-tests.ts — per-adapter
  useTransactionalTests WeakMap and set/get accessors.

Call sites updated to import from the new modules; the wrapper's
tables getter and resetTestAdapterState route through ddl-tracker's
getCreatedTables() / clearDdlTrackers().

Also moves the parseCreateTableColumns unit test alongside the parser
(src/test-adapter.test.ts → src/test-helpers/ddl-tracker.test.ts).

test-adapter.ts: 769 → 547 LOC. Path 3 (slim TestAdapterFixtures of
~30 pure-delegation methods) becomes the natural follow-up.
